modify timeElapsed from int to double

+++ src/main/java/model/entities/Brick.java
@@ -144,7 +144,7 @@ public class Brick extends GameObjectImpl {
      * 
      */
     @Override
-    public void updatePhysics(final int timeElapsed, final GameBoardImpl world) {
+    public void updatePhysics(final double timeElapsed, final GameBoardImpl world) {
         super.getComponentPhysics().update(timeElapsed, this, world);
     }
